refactoring newForm to general form (and adding view & add form logic)
